% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myoglobin_calibration.R
\name{myoglobin_calibration_ratios}
\alias{myoglobin_calibration_ratios}
\title{Extract summarized myoglobin calibration ratios}
\usage{
myoglobin_calibration_ratios(x)
}
\arguments{
\item{x}{A \code{myoglobin_calibration} object.}
}
\value{
A data frame containing the three summarized K/S ratios for each
prepared myoglobin form.
}
\description{
Extract summarized myoglobin calibration ratios
}
