% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{lab_to_hex}
\alias{lab_to_hex}
\title{Convert LAB color values to hexadecimal colors}
\usage{
lab_to_hex(l, a, b, fixup = FALSE)
}
\arguments{
\item{l}{L* values}

\item{a}{a* values}

\item{b}{b* values}

\item{fixup}{Logical. If \code{TRUE}, out-of-gamut colors are corrected to the
closest displayable sRGB color. Defaults to \code{FALSE}.}
}
\value{
A character vector of hexadecimal colors
}
\description{
Convert LAB color values to hexadecimal colors
}
